requirements.txt
httpx>=0.27.0
trafilatura>=1.0.0
readability-lxml>=0.8.0
pypdf>=3.0.0
playwright>=1.40.0
playwright-stealth>=1.0.0
selenium>=4.15.0
webdriver-manager>=4.0.1
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0